+++ src/option.h
@@ -1,0 +1,9 @@
+#include <string>
+
+struct Option {
+  static std::string outputFile;
+  static bool dumpTokens;
+  static bool dumpAST;
+  static bool dumpIR;
+  static bool dumpRegisterAllocation;
+};
